feat: Orphan operations, fixes #9
