+++ src/state/Orphan/interface.ts
@@ -1,0 +1,4 @@
+
+export interface IOrphanValue {
+    payload: any
+}
