+++ src/state/PersistentStateHandler/interface.ts
@@ -1,7 +1,3 @@
-export interface IStateHandler<Type> {
-    get: () => Type
-    set: (value: Type) => void
-}
 
 export interface IStateHandlerAsync {
     get: () => Promise<any>
