+++ src/state/StateHandler/interface.ts
@@ -1,0 +1,4 @@
+export interface IStateHandler<Type> {
+    get: () => Type
+    set: (value: Type) => void
+}
